# autojson/swagger-output.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "Innoart API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "servers": [
    {
      "url": "http://localhost:9000/",
      "description": "Local server"
    },
    {
      "url": "http://localhost:4500/",
      "description": "Testing server"
    },
    {
      "url": "http://petstore.swagger.io/v2/swagger.json",
      "description": "Production server"
    }
  ],
  "tags": [
    {
      "name": "Book",
      "description": "Endpoints"
    }
  ],
  "paths": {
    "/books": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/home": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/check/{token}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/definitions": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Add a user",
            "schema": {
              "$ref": "#/components/schemas/Pet"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Parents": {
        "type": "object",
        "properties": {
          "father": {
            "type": "string",
            "example": "Simon Doe"
          },
          "mother": {
            "type": "string",
            "example": "Marie Doe"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "format": {
                    "type": "string",
                    "example": "int64"
                  }
                }
              },
              "username": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "firstName": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "lastName": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "email": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "password": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "phone": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "userStatus": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "format": {
                    "type": "string",
                    "example": "int32"
                  },
                  "description": {
                    "type": "string",
                    "example": "User Status"
                  }
                }
              }
            }
          },
          "xml": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "User"
              }
            }
          }
        }
      },
      "AddUser": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Karthik"
          },
          "age": {
            "type": "number",
            "example": 29
          },
          "about": {
            "type": "string",
            "example": ""
          }
        },
        "required": [
          "name",
          "age"
        ]
      },
      "Order2": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "id1": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "format": {
                    "type": "string",
                    "example": "int64"
                  }
                }
              },
              "petId": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "format": {
                    "type": "string",
                    "example": "int64"
                  }
                }
              },
              "quantity": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "format": {
                    "type": "string",
                    "example": "int32"
                  }
                }
              },
              "shipDate": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              },
              "status": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Order Status"
                  },
                  "enum": {
                    "type": "array",
                    "example": [
                      "placed",
                      "approved",
                      "delivered"
                    ],
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "complete": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "boolean"
                  },
                  "default": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          },
          "xml": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "Order"
              }
            }
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "format": {
                    "type": "string",
                    "example": "int64"
                  }
                }
              },
              "name": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          },
          "xml": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "Category"
              }
            }
          }
        }
      },
      "Tag": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "format": {
                    "type": "string",
                    "example": "int64"
                  }
                }
              },
              "name": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          },
          "xml": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "Tag"
              }
            }
          }
        }
      },
      "Pet": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "required": {
            "type": "array",
            "example": [
              "name",
              "photoUrls"
            ],
            "items": {
              "type": "string"
            }
          },
          "properties": {
            "type": "object",
            "properties": {
              "id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "format": {
                    "type": "string",
                    "example": "int64"
                  }
                }
              },
              "category": {
                "$ref": "#/components/schemas/Category"
              },
              "name": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "doggie"
                  }
                }
              },
              "photoUrls": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "array"
                  },
                  "xml": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "photoUrl"
                      },
                      "wrapped": {
                        "type": "boolean",
                        "example": true
                      }
                    }
                  },
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  }
                }
              },
              "tags": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "array"
                  },
                  "xml": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "tag"
                      },
                      "wrapped": {
                        "type": "boolean",
                        "example": true
                      }
                    }
                  },
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              },
              "status": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "pet status in the store"
                  },
                  "enum": {
                    "type": "array",
                    "example": [
                      "available",
                      "pending",
                      "sold"
                    ],
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "xml": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "Pet"
              }
            }
          }
        }
      },
      "ApiResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "code": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "format": {
                    "type": "string",
                    "example": "int32"
                  }
                }
              },
              "type": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "message": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "name": "Authorization"
      }
    }
  }
}